+++ src/MyMediaLite/Correlation/Cooccurrence.cs
@@ -51,7 +51,7 @@ namespace MyMediaLite.Correlation
 		}
 
 		///
-		protected override float ComputeCorrelationFromOverlap(float overlap, int count_x, int count_y)
+		protected override float ComputeCorrelationFromOverlap(float overlap, float count_x, float count_y)
 		{
 			return overlap;
 		}
